+++ WinPath.Tests/BackupCommandTests.cs
@@ -154,6 +154,17 @@ namespace WinPath.Tests
 
         [Fact]
         [SupportedOSPlatform("windows")]
+        public void CreateBackupWithoutUserAndSystemFlags()
+        {
+            Program.Main(new string[]
+            {
+                "backup",
+                "create"
+            });
+        }
+
+        [Fact]
+        [SupportedOSPlatform("windows")]
         public void CreateUserBackup()
         {
             Program.Main(
@@ -213,6 +224,18 @@ namespace WinPath.Tests
         }
 
         [Fact]
+        [SupportedOSPlatform("windows")]
+        public void CreateNeitherUserOrSystemBackup()
+        {
+            Program.Main(
+                new string[]
+                {
+                    "backup",
+                    "create"
+                }
+            );
+        }
+
         [SupportedOSPlatform("windows")]
         public void CreateUserBackupInASpecifiedDirectory()
         {