+++ WinPath.Tests/UpdateTests.cs
@@ -78,18 +78,18 @@ namespace WinPath.Tests
         {
             string architecture = Update.GetArchitecture(System.Runtime.InteropServices.RuntimeInformation.OSArchitecture).ToLower();
             Update update = new Update(true, false, (architecture == "x64" || architecture == "x86"));
-            
+
             var releases = update.GetReleases();
             var release = update.FilterRelease(releases);
 
             output.WriteLine(
-                "Release name: " + release.ReleaseName + "\n"
-              + "Release tag: " + release.TagName + "\n"
-              + "Is Prerelease: " + release.IsPrerelease
+                "Release name: " + release?.ReleaseName + "\n"
+              + "Release tag: " + release?.TagName + "\n"
+              + "Is Prerelease: " + release?.IsPrerelease
             );
 
-            Assert.True(release.IsPrerelease);
-            Assert.NotEmpty(release.Assets);
+            Assert.True(release?.IsPrerelease);
+            Assert.NotEmpty(release?.Assets);
         }
 
         [Fact]
@@ -102,13 +102,13 @@ namespace WinPath.Tests
             var release = update.FilterRelease(releases);
 
             output.WriteLine(
-                "Release name: " + release.ReleaseName + "\n"
-              + "Release tag: " + release.TagName + "\n"
-              + "Is Prerelease: " + release.IsPrerelease
+                "Release name: " + release?.ReleaseName + "\n"
+              + "Release tag: " + release?.TagName + "\n"
+              + "Is Prerelease: " + release?.IsPrerelease
             );
 
             if (!releases.Where(release => release.IsPrerelease == true).Any()) // If all releases are NOT prereleases.
-                Assert.False(release.IsPrerelease);                             // Then check if this one isn't a prerelease.
+                Assert.False(release?.IsPrerelease);                             // Then check if this one isn't a prerelease.
             else Assert.True(true);                                             // Else simply pass.
         }
 
@@ -117,20 +117,20 @@ namespace WinPath.Tests
         {
             string architecture = Update.GetArchitecture(System.Runtime.InteropServices.RuntimeInformation.OSArchitecture).ToLower();
             Update update = new Update(true, false, (architecture == "x64" || architecture == "x86"));
-            
+
             var releases = update.GetReleases();
             var release = update.FilterRelease(releases);
 
             output.WriteLine(
-                "Release name: " + release.ReleaseName + "\n"
-              + "Release tag: " + release.TagName + "\n"
-              + "Is Prerelease: " + release.IsPrerelease
+                "Release name: " + release?.ReleaseName + "\n"
+              + "Release tag: " + release?.TagName + "\n"
+              + "Is Prerelease: " + release?.IsPrerelease
             );
 
-            if (!release.ReleaseName.Contains("Test release")) // If the release is not a test release,
-                Assert.NotEmpty(release.Assets);               // check if it's not empty.
-            else                                               //
-                Assert.True(true);                             // Else simply pass.
+            if (!(bool)release?.ReleaseName.Contains("Test release")) // If the release is not a test release,
+                Assert.NotEmpty(release?.Assets);                     // check if it's not empty.
+            else                                                      //
+                Assert.True(true);                                    // Else simply pass.
         }
 
         [Fact]
@@ -141,11 +141,49 @@ namespace WinPath.Tests
 
             var releases = update.GetReleases();
             var release = update.FilterRelease(releases);
-            var assetForProcess = (Asset?)update.GetAssetForProcess(release);
+            var assetForProcess = (Asset?)update.GetAssetForProcess((Release)release);
 
             output.WriteLine("Executable name: " + assetForProcess?.ExecutableName ?? "None");
 
             Assert.True(assetForProcess is not null);
+        }
+
+        // TODO: Remove starting from v1.0.0.
+        /// <summary>
+        /// Runs the update command with prerelease flag.
+        /// </summary>
+        /// <remarks>
+        /// This command runs <c>winpath update</c>, simply this command.
+        /// This is used to test whether it picks a prerelease or not
+        /// since the prerelease flag is not used.
+        /// </remarks>
+        [Fact]
+        [System.Runtime.Versioning.SupportedOSPlatform("windows")]
+        public void GetNullRelease()
+        {
+            Program.Main(
+                new string[] { "update" }
+            );
+        }
+
+        [Fact]
+        [System.Runtime.Versioning.SupportedOSPlatform("windows")]
+        public void UpdatePathWithoutEndingSemicolon()
+        {
+            string initialPath = Environment.GetEnvironmentVariable("Path", EnvironmentVariableTarget.User);
+            if (initialPath.EndsWith(";"))
+                initialPath = initialPath.TrimEnd(';');
+
+            Environment.SetEnvironmentVariable("Path", initialPath, EnvironmentVariableTarget.User);
+
+            Program.Main(
+                new string[]
+                {
+                    "update",
+                    "--prerelease",
+                    "-y"
+                }
+            );
         }
     }
 }